highlight: extract low-level highlight logic from syntax, ui
